Art progress II
I edited a few of the data files to match the new media files.
Here are the new art additions:
- electric beams
- main engine
- reverse engines
- damaged ship
I edited all media files accordingly and added more sprites to the
planet animation making it smoother. Additionally, I in class Animation
I updated the random animation for the beams.